integrations: add notmuch section
Added a section on notmuch integration in the wiki, based on a working
setup with protonmail, with various scenarios for syncing with mbsync.
I thought it would be nice to have different approaches included so that
people can get ideas on how to go about it in a way that suits their
particular setup.

Signed-off-by: Kenneth Flak <kennethflak@protonmail.com>