binds: Open msgs in default Postpone folder in edit-mode
Add a default keybind to open messages in the default Postpone folder
("Drafts") in edit-mode.

A frequent question on IRC is "how do I edit a draft?". When a user
selects a Draft and presses <Enter>, the intent is usually to edit that
draft. This commit makes this the default behaviour for the default
Postpone folder.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>