config: add columns based index format
The index-format option comes from mutt and is neither user friendly,
nor intuitive. Introduce a new way of configuring the message list
contents. Replace index-format with multiple settings to make everything
more intuitive. Reuse the table widget added in the previous commit.

index-columns
	Comma-separated list of column names followed by optional
	alignment and width specifiers.

column-separator
	String separator between columns.

column-$name
	One setting for every name defined in index-columns. This
	supports golang text/template syntax and allows access to the
	same message information than before and much more.

When index-format is still defined in aerc.conf (which will most likely
happen when users will update after this patch), convert it to the new
index-columns + column-$name and column-separator system and a warning
is displayed on startup so that users are aware that they need to update
their config.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Tim Culverhouse <tim@timculverhouse.com>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>